feat: Update Auto-stable deploy strategy logic (#147)
Changed the logic to use a queue-based approach
instead of creating a pending PipelineRun.
The operator now puts CDStageDeploy objects into a queue,
processes them sequentially based on their creation timestamps,
and creates a PipelineRun only after the previous one has completed.
